services/pr_merge_agent/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.10.4
pydantic-settings==2.7.1
requests==2.32.3
clickhouse-connect==0.8.10
jinja2==3.1.5
python-dotenv==1.0.1
neo4j==5.26.0
